encoding and id

+++ BaseObjectModel/src/com/technoserv/bio/kernel/objectmodel/AbstractObject.java
@@ -3,6 +3,6 @@ package com.technoserv.bio.kernel.objectmodel;
 import java.util.Date;
 
 public abstract class AbstractObject {
-	Integer id; // уникальный идентификатор объекта
+	Long id; // уникальный идентификатор объекта
 	Date objectDate; // дата создания объекта
 }
